% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_elbo}
\alias{compare_elbo}
\title{Compare different trained \code{\link{MOFA}} objects in terms of the final value of the ELBO statistics and number of inferred factors}
\usage{
compare_elbo(models)
}
\arguments{
\item{models}{a list containing \code{\link{MOFA}} objects.}
}
\description{
Different objects of \code{\link{MOFA}} are compared in terms of the final value of the ELBO statistics.
For model selection the model with the highest ELBO value is selected.
}
