% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mofa.R
\name{.create_smofa_from_seurat}
\alias{.create_smofa_from_seurat}
\alias{.create_mofa_from_seurat}
\title{create a MOFA object}
\usage{
.create_mofa_from_seurat(
  seurat,
  groups,
  assays = "RNA",
  slot = "data",
  features = NULL,
  save_metadata = FALSE
)
}
\arguments{
\item{seurat}{Seurat object}

\item{groups}{a string specifying column name of the samples metadata to use it as a group variable or character vector with group assignment for every sample}

\item{assays}{assays to use for the MOFA model, default is RNA}

\item{slot}{assay slot to be used such as scale.data or data}

\item{features}{a list with vectors, which are used to subset features, with names corresponding to assays; a vector can be provided when only one assay is used}
}
\value{
Returns an untrained \code{\link{MOFA}} object
}
\description{
Method to create a \code{\link{MOFA}} object
}
\keyword{internal}
