% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{factors}
\alias{factors}
\alias{factors<-}
\alias{factors,MOFA-method}
\alias{factors<-,MOFA,vector-method}
\title{factors: set and retrieve factor names}
\usage{
factors(object)

factors(object) <- value

\S4method{factors}{MOFA}(object)

\S4method{factors}{MOFA,vector}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{value}{a character vector of factor names}
}
\value{
character vector with the factor names
}
\description{
factors: set and retrieve factor names
}
