% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{features}
\alias{features}
\alias{features<-}
\alias{features,MOFA-method}
\alias{features<-,MOFA,list-method}
\alias{features_metadata,MOFA-method}
\title{features: set and retrieve feature names}
\usage{
features(object)

features(object) <- value

\S4method{features}{MOFA}(object)

\S4method{features}{MOFA,list}(object) <- value

\S4method{features_metadata}{MOFA}(object)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{value}{list of character vectors with the feature names for every view}

\item{object}{a \code{\link{MOFA}} object.}
}
\value{
list of character vectors with the feature names for each view

a data frame with sample metadata
}
\description{
features: set and retrieve feature names
}
