% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{features_metadata}
\alias{features_metadata}
\alias{features_metadata<-}
\alias{features_metadata<-,MOFA,data.frame-method}
\title{features_metadata: set and retrieve feature metadata}
\usage{
features_metadata(object, format = "default")

features_metadata(object) <- value

\S4method{features_metadata}{MOFA,data.frame}(object) <- value
}
\arguments{
\item{value}{data frame with feature information, it at least must contain the columns \code{feature_name} and \code{view_name}}
}
\description{
features_metadata: set and retrieve feature metadata
}
