% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{filter_imputation}
\alias{filter_imputation}
\title{Filter imputation results}
\usage{
filter_imputation(object, cutoff = c("stringent", "medium", "lenient"))
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{cutoff}{one of: "stringent", "medium", "lenient"}
}
\value{
a \code{\link{MOFA}} object where the imputed values that do not pass quality threshold have been masked with NA.
}
\description{
This function uses the latent factors and the loadings to impute missing values.
}
\details{
(TO FINISH) MOFA is a probabilistic model and calculates uncertainity for the parameter estimates.
This function can be used to filter the imputation results based on different thresholds of imputation confidence.
The confidence score is calculated as the mean divided by the standard deviation, which provides a signal-to-noise ratio.
}
