% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mofa.R
\name{get_default_data_options}
\alias{get_default_data_options}
\title{Get default data options}
\usage{
get_default_data_options(object)
}
\arguments{
\item{object}{an untrained \code{\link{MOFA}} object}
}
\value{
Returns a list with the default data options.
}
\description{
Function to obtain the default data options.
}
\details{
The data options are the following: \cr
\itemize{
 \item{\strong{scale_views}:}{ logical indicating whether to scale views to have the same unit variance. 
 As long as the scale differences between the data sets is not too high, this is not required. Default is FALSE.}
}
}
