% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mofa.R
\name{get_default_smooth_options}
\alias{get_default_smooth_options}
\title{Get default options for smooth covariates}
\usage{
get_default_smooth_options(object)
}
\arguments{
\item{object}{an untrained \code{\link{MOFA}} object}
}
\value{
Returns a list with default options for the smooth covariate(s) functionality.
}
\description{
Function to obtain the default options for the usage of smooth covariates
}
\details{
The options are the following: \cr
\itemize{
 \item{\strong{scale_cov}:}
 \item{\strong{start_opt}:}
 \item{\strong{n_grid}:}
 \item{\strong{opt_freq}:}
 \item{\strong{sparseGP}:}
 \item{\strong{n_inducing}:}
 \item{\strong{warping}:}
 \item{\strong{warping_freq}:}
 \item{\strong{warping_ref}:}
 \item{\strong{warping_open_begin}:}
 \item{\strong{warping_open_end}:}
 \item{\strong{model_groups}:}
}
}
