% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_group_kernel}
\alias{get_group_kernel}
\title{Get group covariance matrix}
\usage{
get_group_kernel(object)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}
}
\description{
Extract the inferred group-group covariance matrix per factor
}
\details{
This can be used only if covariates are passed to the MOFAobject upon creation and GP_factors is set to True.
}
