% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_imputed_data}
\alias{get_imputed_data}
\title{Get imputed data}
\usage{
get_imputed_data(
  object,
  views = "all",
  groups = "all",
  features = "all",
  as.data.frame = FALSE,
  add_intercept = TRUE,
  only_mean = TRUE
)
}
\arguments{
\item{object}{a trained \code{\link{MOFA}} object.}

\item{views}{character vector with the view name(s), or numeric vector with the view index(es). 
Default is "all".}

\item{groups}{character vector with the group name(s), or numeric vector with the group index(es).
Default is "all".}

\item{features}{list of character vectors with the feature names or list of numeric vectors with the feature indices. 
Default is "all".}

\item{as.data.frame}{logical indicating whether to return a long-formatted data frame instead of a list of matrices. 
Default is \code{FALSE}.}

\item{only_mean}{logical indicating whether to return only the point estimates for the imputation. 
If FALSE, it also retrieves the variance (only if it has been previously calculated).}
}
\value{
TO FINISH
}
\description{
Function to get the imputed data. It requires the previous use of the \code{\link{impute}} method.
}
\details{
TO FINISH
}
