% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_interpolated_factors}
\alias{get_interpolated_factors}
\title{Get interpolated factor values}
\usage{
get_interpolated_factors(object, as.data.frame = FALSE)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}
}
\description{
Extract the interpolated factor values
}
\details{
This can be used only if covariates are passed to the object upon creation, GP_factors is set to True and new covariates were passed for interpolation.
}
