% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{groups}
\alias{groups}
\alias{groups<-}
\alias{groups,MOFA-method}
\alias{groups<-,MOFA,character-method}
\title{groups: set and retrieve group names}
\usage{
groups(object)

groups(object) <- value

\S4method{groups}{MOFA}(object)

\S4method{groups}{MOFA,character}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{value}{character vector with the names for each group}
}
\value{
character vector with the names for each sample group
}
\description{
groups: set and retrieve group names
}
