% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSEA.R
\name{heatmap_FSEA}
\alias{heatmap_FSEA}
\title{Heatmap of Feature Set Enrichment Analysis results}
\usage{
heatmap_FSEA(fsea.out, factors = "all", threshold = 0.05,
  max.pathways = 25, log = TRUE, ...)
}
\arguments{
\item{fsea.out}{output of \link{FeatureSetEnrichmentAnalysis} function}

\item{factors}{: to plot only a subset of given factors}

\item{threshold}{p.value threshold to filter out unsignificant feature sets. Default is 0.05.}

\item{max.pathways}{maximum number of enriched pathways to display (keeping pathways with highest pvalues reached for some factor)}

\item{log}{boolean indicating whether to plot the log of the p.values.}

\item{...}{extra arguments to be passed to \link{pheatmap} function}
}
\description{
this method generates a heatmap with the adjusted p.values that result from the the feature set enrichment analysis. Rows are feature sets and columns are factors.
}
\details{
fill this
}
