% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute missing values from a fitted MOFA}
\usage{
impute(object, views = "all", groups = "all", factors = "all",
  type = c("inRange", "response", "link"))
}
\arguments{
\item{object}{a \code{\link{BioFAM}} object.}

\item{views}{character vector with the view name(s), or numeric vector with view index(es).}

\item{groups}{character vector with the group name(s), or numeric vector with group index(es).}

\item{factors}{character vector with the factor names, or numeric vector with the factor index(es).}

\item{type}{type of imputation. 
\itemize{
\item \strong{response}:{gives mean for gaussian and poisson and probabilities for bernoulli.}
\item \strong{link}: {gives the linear predictions.}
\item \strong{inRange}: {rounds the fitted values from "terms" for integer-valued distributions to the next integer (default).}
}}
}
\description{
This function uses the latent factors and the loadings to impute missing values.
}
\details{
matrix factorization models generate a denoised and condensed low-dimensional representation of the data which capture the main sources of heterogeneity of the data. 
These representation can be used to do predictions using the equation \code{Y = WX}. For more details read the supplementary methods of the manuscript. \cr
This method fills the \code{ImputedData} slot by replacing the missing values in the input data with the model predictions.
}
