% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mofa2.R
\name{mofa2_from_mae}
\alias{mofa2_from_mae}
\title{One-liner wrapper to create a MOFA2 object ready for training}
\usage{
mofa2_from_mae(mae, assay.names, groups = NULL, extract_metadata = FALSE, ...)
}
\arguments{
\item{mae}{a \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment}}}

\item{assay.names}{List of assays to include in MOFA model (required).}

\item{groups}{One of the variable names of coldata from which groups should be derived. Default is NULL, only relevant when using the multi-group framework.}

\item{extract_metadata}{Boolean specifying whether sample metadata should be propagated from the MultiAssayExperiment to the MOFA2 object  (default is FALSE).}

\item{...}{additional parameters for MOFA, named as the options of \code{\link[prepare_mofa]{prepare_mofa}}.}
}
\value{
Returns an untrained \code{\link[MOFA2:MOFA]{MOFA}} with specified options
filled in the corresponding slots
}
\description{
\code{mofa2_from_mae} generates a MOFA2 model from a MultiAssayExperiment with default options that is ready for training.
This is a one-line wrapper that combines the use of \code{\link[create_mofa_from_MultiAssayExperiment]{create_mofa_from_MultiAssayExperiment}} followed by \code{\link[prepare_mofa]{prepare_mofa}}.
}
\examples{
# Load package and import dataset
library(mia)
data("HintikkaXOData", package = "mia")
mae <- HintikkaXOData

# Prepare basic model with selected assays
prep_model <- mofa2_from_mae(mae, assay.names = c("counts", "nmr", "signals"))

# Specify grouping variable and extract metadata
prep_model <- mofa2_from_mae(mae, assay.names = c("counts", "nmr", "signals"),
                    groups = "Diet", extract_metadata = TRUE)

# Modify MOFA options with corresponding arguments
prep_model <- mofa2_from_mae(mae, assay.names = c("counts", "nmr", "signals"),
                    num_factors = 5)
}
