% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{\%>\%}
\alias{\%>\%}
\title{Re-exporting the pipe operator
See \code{magrittr::\link[magrittr]{\%>\%}} for details.}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{see \code{magrittr::\link[magrittr]{\%>\%}}}

\item{rhs}{see \code{magrittr::\link[magrittr]{\%>\%}}}
}
\value{
depending on lhs and rhs
}
\description{
Re-exporting the pipe operator
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
