% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefisto.R
\name{plot_alignment}
\alias{plot_alignment}
\title{Plot covariate alignment acorss groups}
\usage{
plot_alignment(object)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object using MEFISTO with warping}
}
\value{
ggplot object showing the alignment
}
\description{
Function to plot the alignment learnt by MEFISTO for the 
covariate values between different groups
}
\details{
This function requires the functional MEFISTO framework to be used in training. 
Use \code{set_covariates} and specify mefisto_options when preparing the training using \code{prepare_mofa}.
}
