% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_ascii_data}
\alias{plot_ascii_data}
\title{Visualize the structure of the data in the terminal}
\usage{
plot_ascii_data(object, header = FALSE)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object}

\item{header}{a logical value specifying whether to show the MOFA header.}
}
\description{
A Fancy printing method
}
\details{
This function is helpful to get an overview of the structure of the data as a text output
}
