% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{plot_dimred}
\alias{plot_dimred}
\title{Plot dimensionality reduction based on MOFA factors}
\usage{
plot_dimred(object, method = c("umap", "tsne"), groups = "all",
  show_missing = TRUE, color_by = NULL, shape_by = NULL,
  color_name = NULL, shape_name = NULL, dot_size = 1.5, alpha = 1,
  legend = TRUE, return_data = FALSE)
}
\arguments{
\item{object}{a trained \code{\link{MOFA}} object.}

\item{method}{string indicating which method has been used for non-linear dimensionality reduction (either 'umap' or 'tsne')}

\item{groups}{character vector with the groups names, or numeric vector with the indices of the groups of samples to use, or "all" to use samples from all groups.}

\item{show_missing}{logical indicating whether to include samples for which \code{shape_by} or \code{color_by} is missing}

\item{color_by}{specifies groups or values used to color the samples. This can be either:
(1) a character giving the name of a feature present in the training data.
(2) a character giving the same of a column present in the sample metadata.
(3) a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.}

\item{shape_by}{specifies groups or values used to shape the samples. This can be either:
(1) a character giving the name of a feature present in the training data, 
(2) a character giving the same of a column present in the sample metadata.
(3) a vector of the same length as the number of samples specifying discrete groups.}

\item{color_name}{name for color legend.}

\item{shape_name}{name for shape legend.}

\item{dot_size}{numeric indicating dot size.}

\item{alpha}{numeric indicating dot transparency.}

\item{legend}{logical indicating whether to add legend.}

\item{return_data}{logical indicating whether to return the long data frame to plot instead of plotting}
}
\value{
Returns a \code{ggplot2} object or a long data.frame (if return_data is TRUE)
}
\description{
Plot dimensionality reduction based on MOFA factors
}
\details{
TO-FINISH...
}
