% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors}
\alias{plot_factors}
\title{Scatterplots of two factor values}
\usage{
plot_factors(object, factors = c(1, 2), show_missing = TRUE,
  scale = FALSE, color_by = NULL, shape_by = NULL,
  color_name = NULL, shape_name = NULL, dot_size = 1.5, alpha = 1,
  legend = TRUE, return_data = FALSE)
}
\arguments{
\item{object}{a trained \code{\link{MOFA}} object.}

\item{factors}{a vector of length two with the factors to plot. Factors can be specified either as a characters
using the factor names, or as numeric with the index of the factors}

\item{show_missing}{logical indicating whether to include samples for which \code{shape_by} or \code{color_by} is missing}

\item{color_by}{specifies groups or values used to color the samples. 
This can be either 
the string "group" : in this case, the plot will color samples with respect to the groups they belong to
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups or continuous numeric values.}

\item{shape_by}{specifies groups or values used to shape the samples. 
This can be either
a character giving the name of a feature present in the training data, 
a character giving the same of a covariate (only if using \code{\link{MultiAssayExperiment}} as input), 
or a vector of the same length as the number of samples specifying discrete groups.}

\item{color_name}{name for color legend (usually only used if color_by is not a character itself)}

\item{shape_name}{name for shape legend (usually only used if shape_by is not a character itself)}

\item{return_data}{logical indicating whether to return the data frame to plot instead of plotting}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Scatterplot of the values of two latent factors.
}
\details{
One of the first steps for the annotation of factors is to visualise and group/color them using known covariates such as phenotypic or clinical data.
This method generates a single scatterplot for the combination of two latent factors.
Similar function is
\code{\link{plot_factors}} for doing Beeswarm plots for factors.
}
