% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors_on_cov_2d}
\alias{plot_factors_on_cov_2d}
\title{Plot function values on a 2-dimensional grid given by 2 covariates}
\usage{
plot_factors_on_cov_2d(
  object,
  covariates,
  factors = "all",
  scale = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{object}{a trained \code{\link{SMOFA}} object.}

\item{covariates}{character or numeric vector of length 2 specifying the covariates to plot}

\item{scale}{logical indicating whether to scale factor values.}

\item{return_data}{logical indicating whether to return the data frame to plot instead of plotting}

\item{factor}{character or numeric specifying the factor to plot}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Plot of function values on a 2-dimensional grid given by 2 covariates
}
\details{
To investigate the factors pattern along two covariates (such as in the space defined by spatial covaraites) 
this function an be used to plot a factor's values in this covariate space
}
