% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefisto.R
\name{plot_sharedness}
\alias{plot_sharedness}
\title{Barplot showing the sharedness per factor}
\usage{
plot_sharedness(object, factors = "all", color = "#B8CF87")
}
\arguments{
\item{object}{a trained \code{\link{MOFA}} object.}

\item{factors}{character vector with the factors names, or numeric vector indicating the indices of the factors to use}

\item{color}{for the shared part of the bar}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Barplot indicating a sharedness score (between 0 (non-shared) and 1 (shared)) per factor
}
\details{
The sharedness score is calculated as the distance of the learnt group correlation matrix to the identity matrix
 in terms of the mean absolute distance on the off-diagonal elements.
}
