% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot distribution of feature weights (loadings)}
\usage{
plot_weights(object, view = 1, factors = "all", nfeatures = 10,
  color_by = NULL, shape_by = NULL, abs = FALSE, manual = NULL,
  color_manual = NULL, scale = TRUE, dot_size = 1, text_size = 5,
  legend = TRUE, return_data = FALSE)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{view}{a string with the view name, or an integer with the index of the view.}

\item{factors}{character vector with the factor name(s), or numeric vector with the index of the factor(s).}

\item{nfeatures}{number of top features to label.}

\item{color_by}{specifies groups or values (either discrete or continuous) used to color the dots (features). This can be either: 
\itemize{
\item (default) the string "group": in this case, the plot will color the dots with respect to their predefined groups.
\item a character giving the name of a feature that is present in the input data 
\item a character giving the same of a column in the features metadata slot
\item a vector of the same length as the number of features specifying the value for each feature 
\item a dataframe with two columns: "feature" and "color"
}}

\item{shape_by}{specifies groups or values (only discrete) used to shape the dots (features). This can be either: 
\itemize{
\item (default) the string "group": in this case, the plot will shape the dots with respect to their predefined groups.
\item a character giving the name of a feature that is present in the input data 
\item a character giving the same of a column in the features metadata slot
\item a vector of the same length as the number of features specifying the value for each feature 
\item a dataframe with two columns: "feature" and "shape"
}}

\item{abs}{logical indicating whether to take the absolute value of the weights.}

\item{manual}{A nested list of character vectors with features to be manually labelled (see the example for details).}

\item{color_manual}{a character vector with colors, one for each element of 'manual'}

\item{scale}{logical indicating whether to scale all loadings from -1 to 1 (or from 0 to 1 if abs=TRUE).}

\item{dot_size}{numeric indicating the dot size.}

\item{text_size}{numeric indicating the text size.}

\item{legend}{logical indicating whether to add legend.}

\item{return_data}{logical indicating whether to return the data frame to plot instead of plotting}
}
\description{
An important step to annotate factors is to visualise the corresponding feature loadings. \cr
This function plots all loadings for a given latent factor and view, labeling the top ones. \cr
In contrast, the function \code{\link{plot_top_weights}} displays only the top features with highest loading.
}
\examples{
fm <- load_model("data/simple_model.hdf5")
plot_weights(fm, factors = 1:2, nfeature = 3)
}
