% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot output of Feature Set Enrichment Analysis}
\usage{
plotEnrichment(enrichment.results, factor, alpha = 0.1,
  max.pathways = 25, adjust = TRUE, text_size = 1, dot_size = 5)
}
\arguments{
\item{enrichment.results}{output of \link{runEnrichmentAnalysis} function}

\item{factor}{string with factor name or numeric with factor index}

\item{alpha}{p.value threshold to filter out feature sets}

\item{max.pathways}{maximum number of enriched pathways to display}

\item{adjust}{use adjusted p-values?}

\item{object}{\code{\link{MOFAmodel}} object on which FSEA was performed}
}
\value{
a \code{ggplot2} object
}
\description{
Method to plot the results of the Feature Set Enrichment Analyisis (FSEA)
}
