% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plotEnrichmentBars}
\alias{plotEnrichmentBars}
\title{Barplot of Feature Set Enrichment Analysis results}
\usage{
plotEnrichmentBars(enrichment.results, alpha = 0.05)
}
\arguments{
\item{enrichment.results}{output of \link{runEnrichmentAnalysis} function}

\item{alpha}{FDR threshold for calling enriched feature sets. Default is 0.05}
}
\value{
a \link{ggplot2} object
}
\description{
Method to generate a barplot with the number of enriched feature sets per factor
}
