% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Process count data for use with MOFA}
\usage{
preprocess_data(data, groups = NULL, n_top = NULL, type = NULL)
}
\arguments{
\item{data}{A list of matrices, where each entry corresponds to one view. 
Samples are stored in columns and features in rows. 
Missing values must be filled in prior to creating the MOFA object 
(see for example the CLL tutorial).}

\item{groups}{group information, only relevant when using the multi-group framework.}

\item{n_top}{number of features to keep for each view based on highest variance,
needs to be a vector of same length as data, by default all features are used}

\item{type}{Can be on of 'gaussian', 'count' or 'binary'. 
If not provided, is guessed from the data.}
}
\value{
Returns an data in the same format as provided.
}
\description{
Function to preprocess data for input to MOFA. 
Note that this step is optional and users can provide their already preprocessed
data directly to `create_mofa_object` (recommended).
}
\details{
Selects the most informative features based on variance across samples (within groups) and 
transforms count data to Pearson residuals.
}
\examples{
data <- make_example_data(likelihood = "poisson")$data
pp_data <- preprocess_data(data)
}
