% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{quality_control}
\alias{quality_control}
\title{Quality control}
\usage{
quality_control(object, verbose = FALSE)
}
\arguments{
\item{object}{a trained \code{\link{MOFA}} object.}

\item{verbose}{logical indicating whether to generate a verbose output.}
}
\description{
Function to do quality control on a \code{\link{MOFA}} object.
}
\examples{
# Using an existing trained model on simulated data
file <- system.file("exdata", "model.hdf5", package = "MOFA2")
model <- load_model(file)

# Do quality control
model <- quality_control(model, verbose = TRUE)
}
