% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{samples}
\alias{samples}
\alias{samples<-}
\alias{samples,MOFA-method}
\alias{samples<-,MOFA,list-method}
\title{samples: set and retrieve sample names}
\usage{
samples(object)

samples(object) <- value

\S4method{samples}{MOFA}(object)

\S4method{samples}{MOFA,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{value}{list of character vectors with the sample names for every group}
}
\value{
list of character vectors with the sample names for each group
}
\description{
samples: set and retrieve sample names
}
