% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\name{samples_names}
\alias{samples_names}
\alias{samples_names<-}
\alias{samples_names,MOFA-method}
\alias{samples_names<-,MOFA,list-method}
\title{samples_names: set and retrieve sample names}
\usage{
samples_names(object)

samples_names(object) <- value

\S4method{samples_names}{MOFA}(object)

\S4method{samples_names}{MOFA,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{value}{list of character vectors with the sample names for every group}
}
\value{
list of character vectors with the sample names for each group
}
\description{
samples_names: set and retrieve sample names
}
\examples{
# Using an existing trained model on simulated data
file <- system.file("extdata", "model.hdf5", package = "MOFA2")
model <- load_model(file)
samples_names(model)
}
