% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{select_model}
\alias{select_model}
\title{Select a model from a list of trained \code{\link{MOFA}} objects based on the best ELBO value}
\usage{
select_model(models, plot = FALSE)
}
\arguments{
\item{models}{a list containing \code{\link{MOFA}} objects.}

\item{plot}{boolean indicating whether to show a plot of the ELBO for each model instance}
}
\value{
A \code{\link{MOFA}} object
}
\description{
Different objects of \code{\link{MOFA}} are compared in terms of the final value of the ELBO statistics
and the model with the highest ELBO value is selected.
}
