% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_factors}
\alias{subset_factors}
\title{Subset factors}
\usage{
subset_factors(object, factors)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{factors}{character vector with the factor names, or numeric vector with the index of the factors, 
or logical vector with the views to be kept as TRUE.}
}
\description{
Method to subset (or sort) factors
}
