% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_features}
\alias{subset_features}
\title{Subset features}
\usage{
subset_features(object, view, features)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{view}{character vector with the view name or integer with the view index}

\item{features}{character vector with the sample names, numeric vector with the feature indices 
or logical vector with the samples to be kept as TRUE.}
}
\value{
A \code{\link{MOFA}} object
}
\description{
Method to subset (or sort) features
}
