% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset samples}
\usage{
subset_samples(object, samples)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{samples}{character vector with the sample names, numeric vector with the sample indices or logical vector with the samples to be kept as TRUE.}
}
\description{
Method to subset (or sort) samples
}
