% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{training_stats}
\alias{training_stats}
\title{training_stats: set and retrieve training statistics}
\usage{
training_stats(object)
}
\description{
training_stats: set and retrieve training statistics
}
