% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_methods.R
\docType{methods}
\name{views}
\alias{views}
\alias{views<-}
\alias{views,MOFA-method}
\alias{views<-,MOFA,character-method}
\title{views: set and retrieve view names}
\usage{
views(object)

views(object) <- value

\S4method{views}{MOFA}(object)

\S4method{views}{MOFA,character}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{value}{character vector with the names for each view}
}
\value{
character vector with the names for each view
}
\description{
views: set and retrieve view names
}
