% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_scores.R
\name{calculate_contribution_scores}
\alias{calculate_contribution_scores}
\title{Calculate contribution scores for each data modality in each sample}
\usage{
calculate_contribution_scores(
  object,
  views = "all",
  groups = "all",
  factors = "all",
  scale = TRUE
)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{views}{character vector with the view names, or numeric vector with view indexes. Default is 'all'}

\item{groups}{character vector with the group names, or numeric vector with group indexes. Default is 'all'}

\item{factors}{character vector with the factor names, or numeric vector with the factor indexes. Default is 'all'}

\item{scale}{logical indicating whether to scale the sample-wise variance explained values by the total amount of variance explained per view. 
This is useful when the collection of all factors explain different amounts of variance for each data modality (check with `plot_variance_explained(..., plot_total=T)`)}
}
\value{
adds the contribution scores to the metadata slot (see `samples_metadata(MOFAobject)`) and it also returns a data.frame with the contribution score for each sample and data modality
}
\description{
This function takes a trained MOFA model as input and calculates, **for each sample** a contribution score
that quantifies how much each data modality is by the latent space. In other words, how much each data modality "contributes to the latent space.
}
\details{
TO-FILL
}
\examples{
# Using an existing trained model on simulated data
file <- system.file("extdata", "model.hdf5", package = "MOFA2")
model <- load_model(file)

r2 <- calculate_contribution_scores(model, scale = FALSE)
r2 <- calculate_contribution_scores(model, scale = TRUE)

}
