% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mofa_from_mae.R
\name{create_mofa_from_mae}
\alias{create_mofa_from_mae}
\alias{mofa2}
\alias{mofa2,MultiAssayExperiment-method}
\title{Create MOFA from MAE}
\usage{
mofa2(mae, ...)

\S4method{mofa2}{MultiAssayExperiment}(
  mae,
  experiments,
  assay.types,
  groups = NULL,
  extract_metadata = FALSE,
  ...
)
}
\arguments{
\item{mae}{a \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment}}}

\item{...}{additional parameters for MOFA, named as the options
of \code{\link[MOFA2:prepare_mofa]{prepare_mofa}}.
\itemize{
\item \code{altexps}: Specifies alternative experiments for each experiment.
Only applicable for experiments in \code{SingleCellExperiment} format.
}}

\item{experiments}{Name or index of experiments selected from \code{mae}.}

\item{assay.types}{List of assays to include in MOFA model (default is "counts").}

\item{groups}{One of the variable names of coldata from which groups should be derived (default is NULL).}

\item{extract_metadata}{Boolean specifying whethet metadata should be extracted (default is FALSE).}
}
\value{
Returns an untrained \code{\link[MOFA2:MOFA]{MOFA}} with specified options
filled in the corresponding slots
}
\description{
\code{mofa2} produces a prepared MOFA2 model from a MAE.
}
\examples{
\dontrun{
# Load package and import dataset
library(mia)
data("HintikkaXOData", package = "mia")
mae <- HintikkaXOData

# Prepare basic model with selected assays
prep_model <- mofa2(
    mae,
    experiments = names(experiments(mae)),
    assay.types = c("counts", "nmr", "signals")
    )

# Specify grouping variable and extract metadata
prep_model <- mofa2(
    mae,
    experiments = names(experiments(mae)),
    assay.types = c("counts", "nmr", "signals"),
    groups = "Diet", extract_metadata = TRUE
    )

# Modify MOFA options with corresponding arguments
prep_model <- mofa2(
    mae,
    experiments = names(experiments(mae)),
    assay.types = c("counts", "nmr", "signals"),
    num_factors = 5, stochastic = TRUE
    )
}

# Agglomerate and transform microbiome counts data
mae[[1]] <- agglomerateByRanks(mae[[1]])
mae[[1]] <- transformAssay(
    mae[[1]], method = "clr", pseudocount = 1,
    altexp = altExpNames(mae[[1]])
    )
# Create MOFA model
prep_model <- mofa2(
    mae,
    experiments = c(1, 2),
    altexps = c("Genus", NA),
    assay.types = c("clr", "nmr")
    )

}
