% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mofa.R
\name{create_mofa_from_MultiAssayExperiment}
\alias{create_mofa_from_MultiAssayExperiment}
\title{create a MOFA object from a MultiAssayExperiment object}
\usage{
create_mofa_from_MultiAssayExperiment(
  mae,
  assays = NULL,
  groups = NULL,
  extract_metadata = FALSE
)
}
\arguments{
\item{mae}{a \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment}} object}

\item{assays}{assays to use from the \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment}} object}

\item{groups}{a string specifying column name of the colData to use it as a group variable. 
Alternatively, a character vector with group assignment for every sample.
Default is \code{NULL} (no group structure).}

\item{extract_metadata}{logical indicating whether to incorporate the metadata from the MultiAssayExperiment object into the MOFA object}
}
\value{
Returns an untrained \code{\link{MOFA}} object
}
\description{
Method to create a \code{\link{MOFA}} object from a \code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment}} object
}
\examples{
# Using the miniACC data transform RNASeq assays and select three experiments for MOFA
library(MultiAssayExperiment)
data(miniACC)
# apply log1p to the RNASeq and miRNAseq counts
for (nm in c("RNASeq2GeneNorm", "miRNASeqGene")) {
  assay(miniACC[[nm]], "log1p") <-
    log1p(assay(miniACC[[nm]], "exprs"))
}
# create model with the transformed assay, omitting the 2nd and 4th experiments
model <- create_mofa_from_MultiAssayExperiment(miniACC,
                    assays = list("log1p", NULL, "exprs", NULL,'log1p'),
                    extract_metadata = TRUE)
}
