% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefisto.R
\name{get_default_smooth_options}
\alias{get_default_smooth_options}
\title{Get default options for smooth covariates}
\usage{
get_default_smooth_options(object)
}
\arguments{
\item{object}{an untrained \code{\link{MOFA}} object}
}
\value{
Returns a list with default options for the smooth covariate(s) functionality.
}
\description{
Function to obtain the default options for the usage of smooth covariates
}
\details{
The options are the following: \cr
\itemize{
 \item{\strong{scale_cov}:}
 \item{\strong{start_opt}:}
 \item{\strong{n_grid}:}
 \item{\strong{opt_freq}:}
 \item{\strong{sparseGP}:}
 \item{\strong{n_inducing}:}
 \item{\strong{warping}:}
 \item{\strong{warping_freq}:}
 \item{\strong{warping_ref}:}
 \item{\strong{warping_open_begin}:}
 \item{\strong{warping_open_end}:}
 \item{\strong{model_groups}:}
}
}
\examples{
# generate example data
dd <- make_example_data(sample_cov = seq(0,1,length.out = 200), n_samples = 200,
n_factors = 4, n_features = 200, n_views = 4, lscales = c(0.5, 0.2, 0, 0))
# input data
data <- dd$data
# covariate matrix with samples in columns
time <- dd$sample_cov
rownames(time) <- "time"

# create mofa and set covariates
sm <- create_mofa(data = dd$data)
sm <- set_covariates(sm, covariates = time)

smooth_opt <- get_default_smooth_options(sm)
}
