% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mofa.R
\name{get_default_training_options}
\alias{get_default_training_options}
\title{Get default training options}
\usage{
get_default_training_options(object)
}
\arguments{
\item{object}{an untrained \code{\link{MOFA}}}
}
\value{
Returns a list with default training options
}
\description{
Function to obtain the default training options.
}
\details{
The training options are the following: \cr
\itemize{
 \item{\strong{maxiter}:}{ numeric value indicating the maximum number of iterations. 
 Default is 5000 (a lot). Convergence is assessed using the ELBO statistic.}
 \item{\strong{drop_factor_threshold}:}{ (not functional yet) numeric indicating the threshold on fraction of variance explained to consider a factor inactive and drop it from the model.
 For example, a value of 0.01 implies that factors explaining less than 1\% of variance (in each view) will be dropped.}
 \item{\strong{convergence_mode}:}{ character indicating the convergence criteria, either "slow", "medium" or "fast".}
 \item{\strong{verbose}:}{ logical indicating whether to generate a verbose output.}
 \item{\strong{startELBO}:}{ integer indicating the first iteration to compute the ELBO}
 \item{\strong{freqELBO}:}{ integer indicating the first iteration to compute the ELBO}
 \item{\strong{stochastic}:}{ logical indicating whether to use stochastic variational inference (only required for very big data sets).}
 \item{\strong{gpu_mode}:}{ logical indicating whether to use GPUs (see details).}
 \item{\strong{seed}:}{ numeric indicating the seed for reproducibility (default is 0, random seed).}
}
}
