% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Get dimensions}
\usage{
get_dimensions(object)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}
}
\description{
Extract dimensionalities from the model.
}
\details{
K indicates the number of factors, D indicates the number of features, 
N indicates the (total) number of samples and M indicates the number of views.
}
