% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods.R
\name{get_lengthscales}
\alias{get_lengthscales}
\title{Get lengthscales}
\usage{
get_lengthscales(object)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}
}
\description{
Extract the inferred lengthscale for each factor after model training.
}
\details{
This can be used only if covariates are passed to the MOFAobject upon creation and GP_factors is set to True.
}
