% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_ascii_data}
\alias{plot_ascii_data}
\title{Visualize the structure of the data in the terminal}
\usage{
plot_ascii_data(object, nonzero = FALSE)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object}

\item{nonzero}{a logical value specifying whether to calculate the fraction of non-zero values (non-NA values by default)}
}
\value{
None
}
\description{
A Fancy printing method
}
\details{
This function is helpful to get an overview of the structure of the data as a text output
}
\examples{
# Using an existing trained model
file <- system.file("extdata", "model.hdf5", package = "MOFA2")
model <- load_model(file)
plot_ascii_data(model)
}
