% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot output of gene set Enrichment Analysis}
\usage{
plot_enrichment(
  enrichment.results,
  factor,
  alpha = 0.1,
  max.pathways = 25,
  text_size = 1,
  dot_size = 5
)
}
\arguments{
\item{enrichment.results}{output of \link{run_enrichment} function}

\item{factor}{a string with the factor name or an integer with the factor index}

\item{alpha}{p.value threshold to filter out gene sets}

\item{max.pathways}{maximum number of enriched pathways to display}

\item{text_size}{text size}

\item{dot_size}{dot size}
}
\value{
a \code{ggplot2} object
}
\description{
Method to plot the results of the gene set Enrichment Analyisis
}
\details{
it requires \code{\link{run_enrichment}} to be run beforehand.
}
