% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors_on_cov_2d}
\alias{plot_factors_on_cov_2d}
\title{Plot function values on a 2-dimensional grid given by 2 covariates}
\usage{
plot_factors_on_cov_2d(
  object,
  covariates = NULL,
  factors = "all",
  scale = FALSE,
  return_data = FALSE,
  original = FALSE
)
}
\arguments{
\item{object}{a trained \code{\link{MOFA}} object.}

\item{covariates}{specifies exactly 2 sample covariates to plot against
(1) a character vector with 2 entires which each acan be one of the following
(i) character vector giving the names of two column spresent in the sample covariates or sample metadata.
(ii) a character vector giving the names of two feature present in the training data.
(2) a named list of two vectors of the same length as the number of samples specifying continuous numeric values per sample.
Default is the first two sample covariates in covariates slot}

\item{scale}{logical indicating whether to scale factor values.}

\item{return_data}{logical indicating whether to return the data frame to plot instead of plotting}

\item{original}{logical indicating whether to use covariates before training (unscaled) - only relevant if taken from covariates slot}

\item{factor}{character or numeric specifying the factor to plot}
}
\value{
Returns a \code{ggplot2} object
}
\description{
Plot of function values on a 2-dimensional grid given by 2 covariates
}
\details{
To investigate the factors pattern along two covariates (such as in the space defined by spatial covaraites) 
this function an be used to plot a factor's values in this covariate space
}
