% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_covariates.R
\name{set_covariates}
\alias{set_covariates}
\title{Add covariates to a MOFA model}
\usage{
set_covariates(object, covariates = NULL)
}
\arguments{
\item{object}{an untrained \code{\link{MOFA}}}

\item{covariates}{list of data_options (see \code{\link{get_default_data_options}} details).}
}
\value{
Returns an untrained \code{\link{MOFA}} with covariates filled in the corresponding slots
}
\description{
Function to add  continuous covariate to a \code{\link{MOFA}} object for smooth training (MEFISTO)
}
\details{
To activate the functional MEFISTO framework, specify smooth_options when preparing the training using \code{prepare_mofa}
}
\examples{
#' # Simulate data
dd <- make_example_data(sample_cov = seq(0,1,length.out = 100), n_samples = 100, n_factors = 4)

# Create MOFA object
sm <- create_mofa(data = dd$data)

Add a covariate
sm <- set_covariates(sm, dd$sample_cov)
sm
}
