% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset samples}
\usage{
subset_samples(object, samples)
}
\arguments{
\item{object}{a \code{\link{MOFA}} object.}

\item{samples}{character vector with the sample names or numeric vector with the sample indices.}
}
\value{
A \code{\link{MOFA}} object
}
\description{
Method to subset (or sort) samples
}
\examples{
# Using an existing trained model on simulated data
file <- system.file("extdata", "model.hdf5", package = "MOFA2")
model <- load_model(file)

# (TO-DO) Remove a specific sample from the model (an outlier)
}
